debugging

+++ srcs/parsing_utils.c
@@ -76,16 +76,19 @@ int	init_in(t_node *args, t_cmd *elem, t_input *data)
 			return (1);
 		}
 		elem->in = open("heredoc.tmp", O_RDONLY);
-		error_check(elem->in, "In Open heredoc ", 17, data);
+		if (error_check_nofork(elem->in, "in parsing open ", 16, data))
+			return (1);
 		unlink("heredoc.tmp");
 		return (0);
 	}
 	else if (args->type == REDIR_IN)
 	{
+		if (elem->in)
+			close(elem->in);
 		args = args->next;
 		elem->in_arg = ms_strdup(args->value, data);
 		elem->in = open(elem->in_arg, O_RDONLY);
-		error_check(elem->in, "[ERROR]: Wrong File Descriptor\n", 31, data);
+		error_check_nofork(elem->in, "in parsing open ", 16, data);
 		return (0);
 	}
 	return (1);
@@ -98,7 +101,8 @@ int	init_out(t_node *args, t_cmd *elem, t_input *data)
 		args = args->next;
 		elem->out_arg = ms_strdup(args->value, data);
 		elem->out = open(elem->out_arg, O_WRONLY | O_CREAT | O_APPEND, 00644);
-		error_check(elem->out, "[ERROR]: Wrong File Descriptor\n", 31, data);
+		if (error_check_nofork(elem->out, "in parsing open ", 16, data))
+			return (1);
 		return (0);
 	}
 	else if (args->type == REDIR_OUT)
@@ -106,7 +110,8 @@ int	init_out(t_node *args, t_cmd *elem, t_input *data)
 		args = args->next;
 		elem->out_arg = ms_strdup(args->value, data);
 		elem->out = open(elem->out_arg, O_WRONLY | O_CREAT | O_TRUNC, 00644);
-		error_check(elem->out, "[ERROR]: Wrong File Descriptor\n", 31, data);
+		if (error_check_nofork(elem->out, "in parsing open ", 16, data))
+			return (1);
 		return (0);
 	}
 	return (1);
